requirements.txt
annotated-types==0.7.0
anyio>=4.0.0,<4.6.0
bcrypt==4.0.1
click==8.1.8
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
h11==0.14.0
idna==3.10
iniconfig==2.1.0
packaging==24.2
passlib==1.7.4
pip==25.0.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette>=0.44.0,<0.50.0
setuptools>=65.0.0
typing_extensions==4.12.2
uvicorn>=0.30.0,<0.34.0
